+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/config/RpcEnhancementAutoConfiguration.java
@@ -22,17 +22,17 @@ import java.util.List;
 
 import com.tencent.cloud.polaris.context.ConditionalOnPolarisEnabled;
 import com.tencent.cloud.polaris.context.config.PolarisContextAutoConfiguration;
-import com.tencent.cloud.rpc.enhancement.feign.DefaultEnhancedFeignPluginRunner;
 import com.tencent.cloud.rpc.enhancement.feign.EnhancedFeignBeanPostProcessor;
-import com.tencent.cloud.rpc.enhancement.feign.EnhancedFeignPluginRunner;
-import com.tencent.cloud.rpc.enhancement.feign.plugin.EnhancedFeignPlugin;
-import com.tencent.cloud.rpc.enhancement.feign.plugin.reporter.ExceptionPolarisReporter;
-import com.tencent.cloud.rpc.enhancement.feign.plugin.reporter.SuccessPolarisReporter;
+import com.tencent.cloud.rpc.enhancement.plugin.DefaultEnhancedPluginRunner;
+import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPlugin;
+import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginRunner;
+import com.tencent.cloud.rpc.enhancement.plugin.reporter.ExceptionPolarisReporter;
+import com.tencent.cloud.rpc.enhancement.plugin.reporter.SuccessPolarisReporter;
 import com.tencent.cloud.rpc.enhancement.resttemplate.BlockingLoadBalancerClientAspect;
-import com.tencent.cloud.rpc.enhancement.resttemplate.EnhancedRestTemplateReporter;
-import com.tencent.cloud.rpc.enhancement.scg.EnhancedPolarisHttpClientCustomizer;
-import com.tencent.cloud.rpc.enhancement.scg.EnhancedPolarisHttpHeadersFilter;
+import com.tencent.cloud.rpc.enhancement.resttemplate.EnhancedRestTemplateInterceptor;
+import com.tencent.cloud.rpc.enhancement.scg.EnhancedGatewayGlobalFilter;
 import com.tencent.cloud.rpc.enhancement.webclient.EnhancedWebClientReporter;
+import com.tencent.cloud.rpc.enhancement.webclient.PolarisLoadBalancerClientRequestTransformer;
 import com.tencent.polaris.api.core.ConsumerAPI;
 import com.tencent.polaris.client.api.SDKContext;
 
@@ -46,14 +46,12 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
-import org.springframework.cloud.gateway.config.HttpClientCustomizer;
-import org.springframework.cloud.gateway.filter.headers.HttpHeadersFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.context.annotation.Role;
 import org.springframework.web.client.RestTemplate;
-import org.springframework.web.reactive.function.client.ExchangeFilterFunction;
+import org.springframework.web.reactive.function.client.WebClient;
 
 /**
  * Auto Configuration for Polaris {@link feign.Feign} OR {@link RestTemplate} which can automatically bring in the call
@@ -68,6 +66,26 @@ import org.springframework.web.reactive.function.client.ExchangeFilterFunction;
 @AutoConfigureAfter(PolarisContextAutoConfiguration.class)
 public class RpcEnhancementAutoConfiguration {
 
+	@Bean
+	public EnhancedPluginRunner enhancedFeignPluginRunner(
+			@Autowired(required = false) List<EnhancedPlugin> enhancedPlugins) {
+		return new DefaultEnhancedPluginRunner(enhancedPlugins);
+	}
+
+	@Bean
+	public SuccessPolarisReporter successPolarisReporter(RpcEnhancementReporterProperties properties,
+			SDKContext context,
+			ConsumerAPI consumerAPI) {
+		return new SuccessPolarisReporter(properties, context, consumerAPI);
+	}
+
+	@Bean
+	public ExceptionPolarisReporter exceptionPolarisReporter(RpcEnhancementReporterProperties properties,
+			SDKContext context,
+			ConsumerAPI consumerAPI) {
+		return new ExceptionPolarisReporter(properties, context, consumerAPI);
+	}
+
 	/**
 	 * Configuration for Polaris {@link feign.Feign} which can automatically bring in the call
 	 * results for reporting.
@@ -81,33 +99,10 @@ public class RpcEnhancementAutoConfiguration {
 	protected static class PolarisFeignClientAutoConfiguration {
 
 		@Bean
-		public EnhancedFeignPluginRunner enhancedFeignPluginRunner(
-				@Autowired(required = false) List<EnhancedFeignPlugin> enhancedFeignPlugins) {
-			return new DefaultEnhancedFeignPluginRunner(enhancedFeignPlugins);
-		}
-
-		@Bean
-		public EnhancedFeignBeanPostProcessor polarisFeignBeanPostProcessor(@Lazy EnhancedFeignPluginRunner pluginRunner) {
+		public EnhancedFeignBeanPostProcessor polarisFeignBeanPostProcessor(@Lazy EnhancedPluginRunner pluginRunner) {
 			return new EnhancedFeignBeanPostProcessor(pluginRunner);
 		}
 
-		@Configuration
-		static class PolarisReporterConfig {
-
-			@Bean
-			public SuccessPolarisReporter successPolarisReporter(RpcEnhancementReporterProperties properties,
-					@Autowired(required = false) SDKContext context,
-					@Autowired(required = false) ConsumerAPI consumerAPI) {
-				return new SuccessPolarisReporter(properties, context, consumerAPI);
-			}
-
-			@Bean
-			public ExceptionPolarisReporter exceptionPolarisReporter(RpcEnhancementReporterProperties properties,
-					@Autowired(required = false) SDKContext context,
-					@Autowired(required = false) ConsumerAPI consumerAPI) {
-				return new ExceptionPolarisReporter(properties, context, consumerAPI);
-			}
-		}
 	}
 
 	/**
@@ -125,16 +120,15 @@ public class RpcEnhancementAutoConfiguration {
 		private List<RestTemplate> restTemplates = Collections.emptyList();
 
 		@Bean
-		public EnhancedRestTemplateReporter enhancedRestTemplateReporter(
-				RpcEnhancementReporterProperties properties, SDKContext context, ConsumerAPI consumerAPI) {
-			return new EnhancedRestTemplateReporter(properties, context, consumerAPI);
+		public EnhancedRestTemplateInterceptor enhancedPolarisRestTemplateReporter(@Lazy EnhancedPluginRunner pluginRunner) {
+			return new EnhancedRestTemplateInterceptor(pluginRunner);
 		}
 
 		@Bean
-		public SmartInitializingSingleton setErrorHandlerForRestTemplate(EnhancedRestTemplateReporter reporter) {
+		public SmartInitializingSingleton setPolarisReporterForRestTemplate(EnhancedRestTemplateInterceptor reporter) {
 			return () -> {
 				for (RestTemplate restTemplate : restTemplates) {
-					restTemplate.setErrorHandler(reporter);
+					restTemplate.getInterceptors().add(reporter);
 				}
 			};
 		}
@@ -145,6 +139,7 @@ public class RpcEnhancementAutoConfiguration {
 		public BlockingLoadBalancerClientAspect blockingLoadBalancerClientAspect() {
 			return new BlockingLoadBalancerClientAspect();
 		}
+
 	}
 
 	/**
@@ -155,11 +150,28 @@ public class RpcEnhancementAutoConfiguration {
 	@ConditionalOnClass(name = "org.springframework.web.reactive.function.client.WebClient")
 	protected static class PolarisWebClientAutoConfiguration {
 
+		@Autowired(required = false)
+		private List<WebClient.Builder> webClientBuilder = Collections.emptyList();
+
 		@Bean
-		public ExchangeFilterFunction exchangeFilterFunction(
-				RpcEnhancementReporterProperties properties, SDKContext context, ConsumerAPI consumerAPI) {
-			return new EnhancedWebClientReporter(properties, context, consumerAPI);
+		public EnhancedWebClientReporter exchangeFilterFunction(@Lazy EnhancedPluginRunner pluginRunner) {
+			return new EnhancedWebClientReporter(pluginRunner);
 		}
+
+		@Bean
+		public SmartInitializingSingleton addEnhancedWebClientReporterForWebClient(EnhancedWebClientReporter reporter) {
+			return () -> webClientBuilder.forEach(webClient -> {
+				webClient.filter(reporter);
+			});
+		}
+
+		@Bean
+		@ConditionalOnMissingBean
+		@ConditionalOnClass(name = "org.springframework.cloud.client.loadbalancer.reactive.LoadBalancerClientRequestTransformer")
+		public PolarisLoadBalancerClientRequestTransformer polarisLoadBalancerClientRequestTransformer() {
+			return new PolarisLoadBalancerClientRequestTransformer();
+		}
+
 	}
 
 	/**
@@ -172,17 +184,11 @@ public class RpcEnhancementAutoConfiguration {
 	protected static class PolarisGatewayAutoConfiguration {
 
 		@Bean
-		@ConditionalOnClass(name = {"org.springframework.cloud.gateway.filter.headers.HttpHeadersFilter"})
-		public HttpHeadersFilter enhancedPolarisHttpHeadersFilter() {
-			return new EnhancedPolarisHttpHeadersFilter();
-		}
-
-		@Bean
-		@ConditionalOnClass(name = {"org.springframework.cloud.gateway.config.HttpClientCustomizer"})
-		public HttpClientCustomizer httpClientCustomizer(
-				RpcEnhancementReporterProperties properties, SDKContext context, ConsumerAPI consumerAPI) {
-			return new EnhancedPolarisHttpClientCustomizer(properties, context, consumerAPI);
+		@ConditionalOnClass(name = "org.springframework.cloud.gateway.filter.GlobalFilter")
+		public EnhancedGatewayGlobalFilter enhancedPolarisGatewayReporter(@Lazy EnhancedPluginRunner pluginRunner) {
+			return new EnhancedGatewayGlobalFilter(pluginRunner);
 		}
 
 	}
+
 }