+++ spring-cloud-tencent-rpc-enhancement/src/test/java/com/tencent/cloud/rpc/enhancement/AbstractPolarisReporterAdapterTest.java
@@ -17,32 +17,196 @@
 
 package com.tencent.cloud.rpc.enhancement;
 
+import java.net.SocketTimeoutException;
+import java.net.URI;
+import java.net.URISyntaxException;
+
 import com.tencent.cloud.common.constant.HeaderConstant;
+import com.tencent.cloud.common.constant.RouterConstant;
+import com.tencent.cloud.common.metadata.MetadataContext;
+import com.tencent.cloud.common.util.ApplicationContextAwareUtils;
 import com.tencent.cloud.rpc.enhancement.config.RpcEnhancementReporterProperties;
+import com.tencent.polaris.api.config.Configuration;
+import com.tencent.polaris.api.config.global.APIConfig;
+import com.tencent.polaris.api.config.global.GlobalConfig;
+import com.tencent.polaris.api.plugin.circuitbreaker.ResourceStat;
 import com.tencent.polaris.api.pojo.RetStatus;
-import org.assertj.core.api.Assertions;
+import com.tencent.polaris.api.rpc.ServiceCallResult;
+import com.tencent.polaris.client.api.SDKContext;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.MockedStatic;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+
+import static com.tencent.polaris.test.common.Consts.NAMESPACE_TEST;
+import static com.tencent.polaris.test.common.Consts.SERVICE_PROVIDER;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.mock;
 
 /**
  * Test For {@link AbstractPolarisReporterAdapter}.
  *
  * @author <a href="mailto:iskp.me@gmail.com">Elve.Xu</a> 2022/7/11
  */
+@ExtendWith(MockitoExtension.class)
 public class AbstractPolarisReporterAdapterTest {
+
+	private static MockedStatic<ApplicationContextAwareUtils> mockedApplicationContextAwareUtils;
+	private final RpcEnhancementReporterProperties reporterProperties = new RpcEnhancementReporterProperties();
+	@Mock
+	private SDKContext sdkContext;
+
+	@BeforeAll
+	static void beforeAll() {
+		mockedApplicationContextAwareUtils = Mockito.mockStatic(ApplicationContextAwareUtils.class);
+		mockedApplicationContextAwareUtils.when(() -> ApplicationContextAwareUtils.getProperties(anyString()))
+				.thenReturn("unit-test");
+	}
+
+	@AfterAll
+	static void afterAll() {
+		mockedApplicationContextAwareUtils.close();
+	}
+
+	@BeforeEach
+	void setUp() {
+		MetadataContext.LOCAL_NAMESPACE = NAMESPACE_TEST;
+		MetadataContext.LOCAL_SERVICE = SERVICE_PROVIDER;
+	}
+
+	@Test
+	public void testServiceCallResult() throws URISyntaxException {
+		APIConfig apiConfig = mock(APIConfig.class);
+		doReturn("0.0.0.0").when(apiConfig).getBindIP();
+
+		GlobalConfig globalConfig = mock(GlobalConfig.class);
+		doReturn(apiConfig).when(globalConfig).getAPI();
+
+		Configuration configuration = mock(Configuration.class);
+		doReturn(globalConfig).when(configuration).getGlobal();
+
+		doReturn(configuration).when(sdkContext).getConfig();
+
+		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(reporterProperties, sdkContext);
+
+		ServiceCallResult serviceCallResult;
+
+		HttpHeaders requestHeaders = new HttpHeaders();
+		requestHeaders.add(RouterConstant.ROUTER_LABEL_HEADER, "{\"k1\":\"v1\"}");
+
+		serviceCallResult = adapter.createServiceCallResult(
+				"test",
+				null,
+				null,
+				new URI("http://0.0.0.0/"),
+				requestHeaders,
+				new HttpHeaders(),
+				200,
+				0,
+				null
+		);
+		assertThat(serviceCallResult.getRetStatus()).isEqualTo(RetStatus.RetSuccess);
+
+		serviceCallResult = adapter.createServiceCallResult(
+				"test",
+				null,
+				null,
+				new URI("http://0.0.0.0/"),
+				requestHeaders,
+				new HttpHeaders(),
+				502,
+				0,
+				null
+		);
+		assertThat(serviceCallResult.getRetStatus()).isEqualTo(RetStatus.RetFail);
+
+		serviceCallResult = adapter.createServiceCallResult(
+				"test",
+				null,
+				null,
+				new URI("http://0.0.0.0/"),
+				requestHeaders,
+				null,
+				null,
+				0,
+				new SocketTimeoutException()
+		);
+		assertThat(serviceCallResult.getRetStatus()).isEqualTo(RetStatus.RetTimeout);
+
+		serviceCallResult = adapter.createServiceCallResult(
+				"test",
+				"0.0.0.0",
+				8080,
+				new URI("/"),
+				requestHeaders,
+				new HttpHeaders(),
+				200,
+				0,
+				null
+		);
+		assertThat(serviceCallResult.getRetStatus()).isEqualTo(RetStatus.RetSuccess);
+		assertThat(serviceCallResult.getHost()).isEqualTo("0.0.0.0");
+		assertThat(serviceCallResult.getPort()).isEqualTo(8080);
+	}
+
+	@Test
+	public void testResourceStat() throws URISyntaxException {
+
+		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(reporterProperties, sdkContext);
+
+		ResourceStat resourceStat;
+
+		resourceStat = adapter.createInstanceResourceStat("test",
+				null,
+				null,
+				new URI("http://0.0.0.0/"),
+				200,
+				0,
+				null
+		);
+		assertThat(resourceStat.getRetStatus()).isEqualTo(RetStatus.RetSuccess);
+
+		resourceStat = adapter.createInstanceResourceStat("test",
+				null,
+				null,
+				new URI("http://0.0.0.0/"),
+				null,
+				0,
+				new SocketTimeoutException()
+		);
+		assertThat(resourceStat.getRetStatus()).isEqualTo(RetStatus.RetTimeout);
+
+		resourceStat = adapter.createInstanceResourceStat("test",
+				null,
+				null,
+				new URI("http://0.0.0.0/"),
+				200,
+				0,
+				null
+		);
+		assertThat(resourceStat.getRetStatus()).isEqualTo(RetStatus.RetSuccess);
+	}
 
 	@Test
 	public void testApplyWithDefaultConfig() {
 		RpcEnhancementReporterProperties properties = new RpcEnhancementReporterProperties();
 		// Mock Condition
-		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties);
+		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties, sdkContext);
 
 		// Assert
-		Assertions.assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(true);
+		assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(true);
 	}
 
 	@Test
@@ -52,12 +216,12 @@ public class AbstractPolarisReporterAdapterTest {
 		properties.getStatuses().clear();
 		properties.setIgnoreInternalServerError(false);
 
-		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties);
+		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties, sdkContext);
 
 		// Assert
-		Assertions.assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(true);
-		Assertions.assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(true);
+		assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(true);
+		assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(true);
 	}
 
 	@Test
@@ -67,12 +231,12 @@ public class AbstractPolarisReporterAdapterTest {
 		properties.getStatuses().clear();
 		properties.setIgnoreInternalServerError(true);
 
-		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties);
+		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties, sdkContext);
 
 		// Assert
-		Assertions.assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(true);
+		assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(true);
 	}
 
 	@Test
@@ -82,12 +246,12 @@ public class AbstractPolarisReporterAdapterTest {
 		properties.getStatuses().clear();
 		properties.getSeries().clear();
 
-		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties);
+		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties, sdkContext);
 
 		// Assert
-		Assertions.assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(true);
+		assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(true);
 	}
 
 	@Test
@@ -98,14 +262,15 @@ public class AbstractPolarisReporterAdapterTest {
 		properties.getSeries().clear();
 		properties.getSeries().add(HttpStatus.Series.CLIENT_ERROR);
 
-		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties);
+		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties, sdkContext);
 
 		// Assert
-		Assertions.assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(false);
-		Assertions.assertThat(adapter.apply(HttpStatus.FORBIDDEN)).isEqualTo(true);
+		assertThat(adapter.apply(HttpStatus.OK)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.INTERNAL_SERVER_ERROR)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.BAD_GATEWAY)).isEqualTo(false);
+		assertThat(adapter.apply(HttpStatus.FORBIDDEN)).isEqualTo(true);
 	}
+
 
 	@Test
 	public void testGetRetStatusFromRequest() {
@@ -115,19 +280,19 @@ public class AbstractPolarisReporterAdapterTest {
 		properties.getSeries().clear();
 		properties.getSeries().add(HttpStatus.Series.CLIENT_ERROR);
 
-		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties);
+		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties, sdkContext);
 
 		HttpHeaders headers = new HttpHeaders();
 		RetStatus ret = adapter.getRetStatusFromRequest(headers, RetStatus.RetFail);
-		Assertions.assertThat(ret).isEqualTo(RetStatus.RetFail);
+		assertThat(ret).isEqualTo(RetStatus.RetFail);
 
 		headers.set(HeaderConstant.INTERNAL_CALLEE_RET_STATUS, RetStatus.RetFlowControl.getDesc());
 		ret = adapter.getRetStatusFromRequest(headers, RetStatus.RetFail);
-		Assertions.assertThat(ret).isEqualTo(RetStatus.RetFlowControl);
+		assertThat(ret).isEqualTo(RetStatus.RetFlowControl);
 
 		headers.set(HeaderConstant.INTERNAL_CALLEE_RET_STATUS, RetStatus.RetReject.getDesc());
 		ret = adapter.getRetStatusFromRequest(headers, RetStatus.RetFail);
-		Assertions.assertThat(ret).isEqualTo(RetStatus.RetReject);
+		assertThat(ret).isEqualTo(RetStatus.RetReject);
 	}
 
 	@Test
@@ -138,15 +303,15 @@ public class AbstractPolarisReporterAdapterTest {
 		properties.getSeries().clear();
 		properties.getSeries().add(HttpStatus.Series.CLIENT_ERROR);
 
-		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties);
+		SimplePolarisReporterAdapter adapter = new SimplePolarisReporterAdapter(properties, sdkContext);
 
 		HttpHeaders headers = new HttpHeaders();
 		String ruleName = adapter.getActiveRuleNameFromRequest(headers);
-		Assertions.assertThat(ruleName).isEqualTo("");
+		assertThat(ruleName).isEqualTo("");
 
 		headers.set(HeaderConstant.INTERNAL_ACTIVE_RULE_NAME, "mock_rule");
 		ruleName = adapter.getActiveRuleNameFromRequest(headers);
-		Assertions.assertThat(ruleName).isEqualTo("mock_rule");
+		assertThat(ruleName).isEqualTo("mock_rule");
 	}
 
 	/**
@@ -154,8 +319,8 @@ public class AbstractPolarisReporterAdapterTest {
 	 */
 	public static class SimplePolarisReporterAdapter extends AbstractPolarisReporterAdapter {
 
-		public SimplePolarisReporterAdapter(RpcEnhancementReporterProperties properties) {
-			super(properties);
+		protected SimplePolarisReporterAdapter(RpcEnhancementReporterProperties reportProperties, SDKContext context) {
+			super(reportProperties, context);
 		}
 	}
 }